Update WorkflowSimulationStudyScript_Feb25.R
initialising variables required when analysing covariance (lines 31-38):
actualcovdp()
actualcovdn <- c()
maxcovdp <- c()
mincovdp <- c()
maxcovdn <- c()
mincovdn <- c()
acsimsevolunteer1 <- c() #option to include acsimsevolunteer2 <- c() if desired later in code
acsimspvolunteer1 <- c() #option to include acsimsevolunteer2 <- c() if desired later in code

Line 130. Correct typo in the comment true disease prevalence here is 0.60

Lines 152, 180, 410, 425, 443, 461
Changed the 'True' term in epi.betabuster function to 'greater than' so it is in line with current version of epi.betabuster. Explained in comments that the epi.betabuster
function is generating beta distribution based on expert elicitation that the best estimate for this value is the first term in the function, and we are 95% sure the value is
greater than the second to last term in the function

Lines 579 to 582: Corrected typo here from vol1 to vol1b and vol2 to vol2b - code won't run otherwise